pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "deterministic-rsa-keygen"
version = "1.0"
description = "generate deterministic RSA key pair and perform encrypt and decrypt operations"
readme = "README.md"
authors = ["Mauricio <mtdb.github.n20n0@passinbox.com>"]
license = "Apache Software License"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[tool.pytest.ini_options]
pythonpath = ["."]

[project.urls]
Homepage = "https://github.com/slashpass/deterministic-rsa-keygen"
Issues = "https://github.com/slashpass/deterministic-rsa-keygen/issues"